added empty screen view (appears after deleting last item) recycler view updates in real time (pressing cancel after deleting item brings back the item) updated app icon to facilitate round icons back button on main activity now only closes the app unfinished search view implemented

+++ app/src/main/java/com/example/noted/Adapter.kt
@@ -1,16 +1,24 @@
 package com.example.noted
 
+import android.app.Activity
+import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
 import android.view.LayoutInflater
+import android.view.View
 import android.view.ViewGroup
+import android.widget.TextView
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView.ViewHolder
+import com.example.noted.databinding.ActivityMainBinding
 import com.example.noted.databinding.PreviewBinding
 import java.io.File
 
 class Adapter(private val list: ArrayList<Info>, private val context: Context): RecyclerView.Adapter<Adapter.MyView>() {
     inner class MyView(val itemBinding: PreviewBinding): RecyclerView.ViewHolder(itemBinding.root)
+
+
 
     inner class SwipeToDeleteCallback : ItemTouchHelper.SimpleCallback(
         0,
@@ -26,10 +34,35 @@ class Adapter(private val list: ArrayList<Info>, private val context: Context): 
         }
 
         override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
+            val builder = AlertDialog.Builder(context)
             // Get the position of the swiped item
             val position = viewHolder.adapterPosition
+            builder.setMessage("Are you sure you want to Delete?")
+                .setCancelable(false)
+                .setPositiveButton("Delete") { dialog, id ->
+                    // Delete selected note from database
+                    deleteItem(position)
+                    val intent = Intent(context, MainActivity::class.java)
+                    (context as Activity).finish()
+                    context.overridePendingTransition(0,0)
+                    context.startActivity(intent)
+                    context.overridePendingTransition(0,0)
+                }
+                .setNegativeButton("Cancel") { dialog, id ->
+                    // Dismiss the dialog
+                    dialog.dismiss()
+                    val intent = Intent(context, MainActivity::class.java)
+                    (context as Activity).finish()
+                    context.overridePendingTransition(0,0)
+                    context.startActivity(intent)
+                    context.overridePendingTransition(0,0)
+                }
+            val alert = builder.create()
+            alert.show()
+
+
             // Delete the item from the adapter's data source
-            deleteItem(position)
+            //deleteItem(position)
         }
     }
 
@@ -38,7 +71,6 @@ class Adapter(private val list: ArrayList<Info>, private val context: Context): 
         file.delete()
         list.removeAt(position)
         notifyItemRemoved(position)
-
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MyView {
@@ -63,6 +95,17 @@ class Adapter(private val list: ArrayList<Info>, private val context: Context): 
         }else{
             holder.itemBinding.tvItem.text = list[position].title
         }
+
+        val tv1 = holder.itemView.findViewById<TextView>(R.id.NOTED)
+        val tv2 = holder.itemView.findViewById<TextView>(R.id.textView2)
+
+//        if(list.isEmpty()){
+//            tv1.visibility = View.VISIBLE
+//            tv2.visibility = View.VISIBLE
+//        }else{
+//            tv1.visibility = View.GONE
+//            tv2.visibility = View.GONE
+//        }
 
 
         holder.itemView.setOnClickListener {
